remove gmp_generic variant too -- reduce number of cgo calls
