+++ crypto/modexp/gmp/README.md
@@ -15,9 +15,7 @@ You need to have GMP development libraries installed on your system:
 
 ## Usage
 
-This package provides a GMP-backed implementation for modular exponentiation. The API can be expanded, however right now, the main usage is for the modexp precompile.
-
-There are currently two implementations: generic (using Go wrapper types) and cwrapper (direct C calls).
+This package provides a GMP-backed implementation for modular exponentiation using direct C calls for maximum performance. The API can be expanded, however right now, the main usage is for the modexp precompile.
 
 ### Byte Array Interface (Recommended)
 
@@ -35,25 +33,6 @@ if err != nil {
 // result = 24 (2^10 mod 1000)
 ```
 
-### Direct GMP Interface
-
-```go
-// Create numbers
-base := gmp.NewInt()
-base.SetString("123456789", 10)
-
-exp := gmp.NewInt()
-exp.SetString("987654321", 10)
-
-mod := gmp.NewInt()
-mod.SetString("1000000007", 10)
-
-// Compute base^exp mod mod
-result := gmp.NewInt()
-result.ExpMod(base, exp, mod)
-
-fmt.Printf("Result: %s\n", result)
-```
 
 
 ## Testing
